Synchronize: fetch all references, not only HEAD
Not sure if this is what we want with the current interface, but at the moment only HEAD is fetched, and there seems to be no way to fetch other branches. It would actually be nice to be able to fetch a specific ref (or even sha, in my use-case I know in advance the sha I need).